1月7日
1、年龄：接口、界面：打开直接跳转到当前出生年月所在的地方；
2、我的标签：一级、二级界面；（使用的统一的界面，只是网络请求时，改变接口）
3、附近的人：根据最近上传的位置信息，返回附近的人信息；

+++ NZNYiOS/NavigationVC/搜索/Controller/CYSearchVC.m
@@ -391,7 +391,7 @@
 
 // cell：点击事件
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath{
-    NSLog(@"点击了第 %ld 行 cell",indexPath.row);
+    NSLog(@"点击了第 %ld 行 cell",(long)indexPath.row);
     
     //当离开某行时，让某行的选中状态消失
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
